backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
SQLAlchemy>=2.0.36
PyJWT==2.8.0
bcrypt==4.0.1
httpx==0.25.2
python-multipart==0.0.6
python-decouple==3.8
pydantic[email]>=2.7
email-validator>=2.1
typing_extensions>=4.12.2
